Implementations [ci skip]
- refactor Predicate
- implements ScheduleOncePredicate
